fix: 빌드 에러 잡기
